Clone database setting for test instead of referencing
Currently this sets `settings.DATABASES['default']['TEST'] = settings.DATABASES['default']` when `'CI' in os.environ`. This creates a circular reference. If a test triggers an error inside of a view, django tries to `cleanse_settings` to display it on a debug page which raises `builtins.RuntimeError: maximum recursion depth exceeded` because of the loop. By copying the database settings instead of referencing it, the circular reference is avoided.